refactor: move links retrieval from object to refs
